Consolidate some of the types from the LS.Protocol project into a Workspaces.Protocol project
