+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Protocol/IDelegatedParams.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-using Microsoft.CodeAnalysis.Razor.Workspaces;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
+namespace Microsoft.CodeAnalysis.Razor.Workspaces.Protocol;
 
 /// <summary>
 /// Interface for delegated params that enables sharing of code in RazorCustomMessageTarget
