+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Protocol/RazorLanguageKind.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-namespace Microsoft.CodeAnalysis.Razor.Workspaces;
+namespace Microsoft.CodeAnalysis.Razor.Workspaces.Protocol;
 
 internal enum RazorLanguageKind
 {
